backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
groq==1.0.0
google-generativeai==0.3.2
supabase==2.28.0
python-dotenv==1.0.0
httpx==0.28.1
websockets==15.0.1
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3
